[Prefactor] Move controller to `runtime` package
