+++ docs/concepts/operator.md
@@ -552,6 +552,7 @@ Some controllers may only be instantiated or added later, because they need the 
 
 * [`NetworkPolicy` controller](gardenlet.md#networkpolicy-controller)
 * [`VPA EvictionRequirements` controller](gardenlet.md#vpaevictionrequirements-controller)
+* [`Required Runtime` reconciler](#required-runtime-reconciler)
 * [`Access` controller](#access-controller)
 * [`Virtual-Cluster-Registrar` controller](#virtual-cluster-registrar-controller)
 * [`Gardenlet` controller](#gardenlet-controller)
@@ -572,11 +573,11 @@ Currently, this logic handles the following scenarios:
 - Extension admission deployment for the virtual garden cluster.
 - `ControllerDeployment` and `ControllerRegistration` reconciliation in the virtual garden cluster.
 
-#### [`Required` Reconciler](../../pkg/operator/controller/extension/required)
+#### [`Required Runtime` Reconciler](../../pkg/operator/controller/extension/required/runtime)
 
 This reconciler reacts on events from `BackupBucket`, `DNSRecord` and `Extension` resources.
 Based on these resources and the related `Extension` specification, it is checked if the extension deployment is required in the garden runtime cluster.
-The result is then put into the `RuntimeRequired` condition and added to the `Extension` status.
+The result is then put into the `RequiredRuntime` condition and added to the `Extension` status.
 
 ### [`Access` Controller](../../pkg/operator/controller/virtual/access)
