+++ pkg/operator/controller/extension/required/runtime/reconciler.go
@@ -2,7 +2,7 @@
 //
 // SPDX-License-Identifier: Apache-2.0
 
-package required
+package runtime
 
 import (
 	"context"
